backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PySocks
python-multipart==0.0.6
httpx==0.26.0
aiohttp==3.9.1
websockets==12.0
aioredis==2.0.1
imaplib2==3.6
mail-parser==3.15.0
beautifulsoup4==4.12.2
lxml==4.9.4
python-dateutil==2.8.2
pytest==7.4.4
pytest-asyncio==0.21.1
httpx==0.26.0